chore: Move demo to dir that triggers chromatic
