+++ packages/plan/src/demo/iterationDiff/IterationChangeCard.tsx
@@ -1,5 +1,5 @@
 import { type KeyboardEvent } from 'react';
-import type { IterationChange } from './iterationChanges.ts';
+import type { IterationChange } from './iterationChanges.tsx';
 import { ThreadedCommentCard } from './ThreadedCommentCard.tsx';
 
 export const iterationChangeCardTestIds = {
