+++ packages/plan/src/demo/iterationDiff/IterationDiffStage.tsx
@@ -7,9 +7,9 @@ import { MarkdownPlan } from '../../MarkdownPlan.tsx';
 import { usePlanAppContext } from '../../useAppContext.ts';
 import { InlineChangeThreads } from './InlineChangeThreads.tsx';
 import { IterationAnchoredCommentsSidebar, type ReplyFocusRequest } from './IterationAnchoredCommentsSidebar.tsx';
-import { findIterationChangeElement, isInlineChange } from './iterationChangeAnchors.ts';
+import { findIterationChangeElement, isInlineChange } from './iterationChangeAnchors.tsx';
 import { IterationChangeHighlights } from './IterationChangeHighlights.tsx';
-import type { IterationChange } from './iterationChanges.ts';
+import type { IterationChange } from './iterationChanges.tsx';
 import { IterationChangesSidebar } from './IterationChangesSidebar.tsx';
 import { IterationThreadNavigationBar } from './IterationThreadNavigationBar.tsx';
 import './iterationDiffStyles.css';
